chore: cleanup rollup implementation for packages
