+++ packages/-build-infra/package.json
@@ -14,6 +14,10 @@
     "test:node": "mocha"
   },
   "dependencies": {
+    "broccoli-funnel": "^2.0.2",
+    "broccoli-rollup": "^4.1.1",
+    "broccoli-merge-trees": "^3.0.2",
+    "broccoli-debug": "^0.6.5",
     "babel-plugin-feature-flags": "^0.3.1",
     "babel-plugin-filter-imports": "^3.0.0",
     "babel6-plugin-strip-class-callcheck": "^6.0.0",
@@ -25,6 +29,7 @@
     "ember-cli-version-checker": "^3.1.2",
     "git-repo-info": "^2.0.0",
     "glob": "^7.1.3",
+    "chalk": "^2.4.1",
     "mocha": "^6.1.2",
     "npm-git-info": "^1.0.3",
     "rimraf": "^2.6.2",
