+++ packages/-ember-data/package.json
@@ -29,21 +29,9 @@
     "@ember-data/store": "^3.12.0-canary.1",
     "@ember/ordered-set": "^2.0.3",
     "@glimmer/env": "^0.1.7",
-    "broccoli-debug": "^0.6.5",
-    "broccoli-funnel": "^2.0.1",
-    "broccoli-merge-trees": "^3.0.2",
-    "broccoli-rollup": "^4.1.1",
-    "chalk": "^2.4.1",
     "ember-cli-babel": "^7.8.0",
-    "ember-cli-string-utils": "^1.1.0",
-    "ember-cli-test-info": "^1.0.0",
     "ember-cli-typescript": "^2.0.1",
-    "ember-cli-version-checker": "^3.1.2",
-    "ember-inflector": "^3.0.0",
-    "inflection": "^1.12.0",
-    "resolve": "^1.8.1",
-    "semver": "^6.1.2",
-    "silent-error": "^1.1.1"
+    "ember-inflector": "^3.0.0"
   },
   "devDependencies": {
     "@babel/plugin-transform-typescript": "*",
