+++ packages/adapter/addon/-private/index.js
@@ -2,3 +2,4 @@ export { default as parseResponseHeaders } from './utils/parse-response-headers'
 export { determineBodyPromise } from './utils/determine-body-promise';
 export { serializeQueryParams } from './utils/serialize-query-params';
 export { default as fetch } from './utils/fetch';
+export { default as BuildURLMixin } from './build-url-mixin';
