+++ packages/model/addon/-private/index.ts
@@ -1,0 +1,3 @@
+export { default as attr } from './attr';
+export { default as belongsTo } from './belongs-to';
+export { default as hasMany } from './has-many';
