+++ packages/model/addon/index.ts
@@ -1,4 +1,2 @@
-export { default as attr } from './-private/attr';
-export { default as belongsTo } from './-private/relationships/belongs-to';
-export { default as hasMany } from './-private/relationships/has-many';
+export { attr, belongsTo, hasMany } from './-private';
 export { Model as default } from '@ember-data/store/-private';
