+++ packages/store/package.json
@@ -27,8 +27,7 @@
     "ember-cli-babel": "^7.8.0",
     "ember-cli-path-utils": "^1.0.0",
     "ember-cli-typescript": "^2.0.1",
-    "heimdalljs": "^0.3.0",
-    "silent-error": "^1.1.1"
+    "heimdalljs": "^0.3.0"
   },
   "devDependencies": {
     "@ember/optional-features": "*",
